+++ adapter/spark/src/main/java/com/asml/apa/wta/spark/listener/TaskLevelListener.java
@@ -7,12 +7,13 @@ import java.util.Arrays;
 import java.util.List;
 import java.util.Map;
 import java.util.concurrent.ConcurrentHashMap;
+import java.util.stream.Collectors;
 import lombok.Getter;
-import org.apache.commons.lang3.ArrayUtils;
 import org.apache.spark.SparkContext;
 import org.apache.spark.executor.TaskMetrics;
 import org.apache.spark.resource.ResourceProfile;
 import org.apache.spark.resource.TaskResourceRequest;
+import org.apache.spark.scheduler.SparkListenerJobEnd;
 import org.apache.spark.scheduler.SparkListenerTaskEnd;
 import org.apache.spark.scheduler.TaskInfo;
 import scala.collection.JavaConverters;
@@ -28,9 +29,9 @@ import scala.collection.JavaConverters;
 @Getter
 public class TaskLevelListener extends TaskStageBaseListener {
 
-  private final Map<Integer, List<Task>> stageToTasks = new ConcurrentHashMap<>();
+  private final Map<Long, List<Task>> stageToTasks = new ConcurrentHashMap<>();
 
-  private final Map<Long, Integer> taskToStage = new ConcurrentHashMap<>();
+  private final Map<Long, Long> taskToStage = new ConcurrentHashMap<>();
 
   /**
    * Constructor for the task-level listener.
@@ -45,75 +46,77 @@ public class TaskLevelListener extends TaskStageBaseListener {
   }
 
   /**
-   * This method is called every time a task ends, task-level metrics should be collected here, and added.
+   * Fills in the maps used to determine parent-child relation of Tasks at on application end.
    *
-   * @param taskEnd   SparkListenerTaskEnd The object corresponding to information on task end
+   * @param taskId    Spark Task id
+   * @param stageId   Spark Stage id
+   * @param wtaTask   WTA Task object
+   * @author Pil Kyu Cho
+   * @since 1.0.0
+   */
+  private void fillInParentChildMaps(long taskId, long stageId, Task wtaTask) {
+    taskToStage.put(taskId, stageId);
+    final List<Task> sparkTasks = stageToTasks.computeIfAbsent(stageId, tasks -> new ArrayList<>());
+    sparkTasks.add(wtaTask);
+  }
+
+  /**
+   * This method is called every time a task ends. Task-level metrics are collected, aggregated, and added here.
+   * <p>
+   * Note:
+   * peakExecutionMemory is the peak memory used by internal data structures created during shuffles, aggregations
+   * and joins. The value of this accumulator should be approximately the sum of the peak sizes across all such
+   * data structures created in this task. It is thus only an upper bound of the actual peak memory for the task.
+   * For SQL jobs, this only tracks all unsafe operators and ExternalSort
+   * <p>
+   * Alternative:
+   * final double memoryRequested = curTaskMetrics.peakExecutionMemory();
+   *
+   * @param taskEnd   SparkListenerTaskEnd object corresponding to information on task end
    * @author Henry Page
    * @author Tianchen Qu
+   * @author Pil Kyu Cho
    * @since 1.0.0
    */
   @Override
   public void onTaskEnd(SparkListenerTaskEnd taskEnd) {
     final TaskInfo curTaskInfo = taskEnd.taskInfo();
     final TaskMetrics curTaskMetrics = taskEnd.taskMetrics();
+    assert curTaskMetrics != null;
+
     final long taskId = curTaskInfo.taskId() + 1;
+    final long stageId = taskEnd.stageId() + 1;
+
     final String type = taskEnd.taskType();
-    final long submitTime = curTaskInfo.launchTime();
-    final long runTime = curTaskMetrics.executorRunTime();
+    final long tsSubmit = curTaskInfo.launchTime();
+    final long runtime = curTaskMetrics.executorRunTime();
     final int userId = sparkContext.sparkUser().hashCode();
-    final int stageId = taskEnd.stageId();
-    final long workflowId = stageIdsToJobs.get(taskEnd.stageId() + 1);
-
-    final long[] parents = new long[0];
-    final long[] children = new long[0];
-    taskToStage.put(taskId, stageId);
-
+    final long workflowId = stageToJob.get(stageId);
+    final long diskIoTime = -1L;
     final double diskSpaceRequested = (double) curTaskMetrics.diskBytesSpilled()
         + curTaskMetrics.shuffleWriteMetrics().bytesWritten();
-    /**
-     * alternative:
-     *
-     * final double memoryRequested = curTaskMetrics.peakExecutionMemory();
-     *
-     *  peakExecutionMemory is the peak memory used by internal data structures created during shuffles, aggregations
-     *  and joins.
-     *  The value of this accumulator should be approximately the sum of the peak sizes across all such data structures
-     *  created in this task.
-     *  It is thus only an upper bound of the actual peak memory for the task.
-     *  For SQL jobs, this only tracks all unsafe operators and ExternalSort
-     */
-    final double memoryRequested = -1.0;
-    final String resourceType = "N/A";
-    final double resourceAmountRequested = -1.0;
-    final long diskIoTime = -1L;
-    /**
-     * alternative:
-     *
-     * final long diskIoTime =
-     *         curTaskMetrics.executorDeserializeTime() + curTaskMetrics.resultSerializationTime();
-     *
-     * This is not totally accurate as diskIoTime also includes time accessing the disk in case of a spill.
-     * Also, the deserialization/serialization time in spark is not completely clear if it also considers the time spent
-     * on the operating system side.
-     */
     final long resourceUsed = Math.abs(curTaskInfo.executorId().hashCode());
 
-    // unknown
+    // dummy values
+    final double resourceAmountRequested = -1.0;
     final int submissionSite = -1;
+    final String resourceType = "N/A";
+    final long[] parents = new long[0];
+    final long[] children = new long[0];
     final int groupId = -1;
     final String nfrs = "";
-    final String params = "";
-    final long networkIoTime = -1L;
-
-    final double energyConsumption = -1L;
     final long waitTime = -1L;
+    final String params = "";
+    final double memoryRequested = -1.0;
+    final long networkIoTime = -1L;
+    final double energyConsumption = -1L;
 
     Task task = Task.builder()
         .id(taskId)
         .type(type)
         .submissionSite(submissionSite)
-        .tsSubmit(submitTime)
-        .runtime(runTime)
+        .tsSubmit(tsSubmit)
+        .runtime(runtime)
         .resourceType(resourceType)
         .resourceAmountRequested(resourceAmountRequested)
         .parents(parents)
@@ -131,51 +134,57 @@ public class TaskLevelListener extends TaskStageBaseListener {
         .energyConsumption(energyConsumption)
         .resourceUsed(resourceUsed)
         .build();
-
-    final List<Task> tasks = stageToTasks.get(stageId);
-    if (tasks == null) {
-      List<Task> newTasks = new ArrayList<>();
-      newTasks.add(task);
-      stageToTasks.put(stageId, newTasks);
-    } else {
-      tasks.add(task);
-    }
     this.getProcessedObjects().add(task);
+    fillInParentChildMaps(taskId, stageId, task);
   }
 
   /**
-   * This method set up the parents and children of all tasks as well as the resources used by all tasks.
+   * Sets the parent, child and resource fields for Spark Tasks. This method is called on job end in
+   * {@link JobLevelListener#onJobEnd(SparkListenerJobEnd)} and only sets the Tasks which are
+   * affiliated to the passed jobId.
    *
-   * @param stageListener stage level listener to get resource bindings
+   * @param stageLevelListener    stage level listener to get ConcurrentHashMap containers
+   * @param jobId                 Spark Job id to filter Tasks by
    * @author Tianchen Qu
+   * @author Pil Kyu Cho
    * @since 1.0.0
    */
-  public void setTasks(StageLevelListener stageListener) {
-    final List<Task> tasks = this.getProcessedObjects();
-    for (Task task : tasks) {
-      // parent children fields
-      final int stageId = this.getTaskToStage().get(task.getId());
-      final Integer[] parentStages = stageListener.getStageToParents().get(stageId);
+  public void setTasks(StageLevelListener stageLevelListener, long jobId) {
+    final List<Task> filteredTasks = this.getProcessedObjects().stream()
+        .filter(t -> t.getWorkflowId() == jobId)
+        .collect(Collectors.toList());
+    for (Task task : filteredTasks) {
+      // set parent field: all Tasks in are guaranteed to be in taskToStage
+      final long stageId = this.getTaskToStage().get(task.getId());
+      final Long[] parentStages = stageLevelListener.getStageToParents().get(stageId);
       if (parentStages != null) {
-        final Long[] parents = Arrays.stream(parentStages)
-            .flatMap(parentStage ->
-                Arrays.stream(stageToTasks.getOrDefault(parentStage, new ArrayList<>()).stream()
+        final long[] parents = Arrays.stream(parentStages)
+            .flatMap(parentStageId -> Arrays.stream(
+                this.getStageToTasks().getOrDefault(parentStageId, new ArrayList<>()).stream()
                     .map(Task::getId)
                     .toArray(Long[]::new)))
-            .toArray(Long[]::new);
-        task.setParents(ArrayUtils.toPrimitive(parents));
+            .mapToLong(Long::longValue)
+            .toArray();
+        task.setParents(parents);
       }
 
-      List<Integer> childrenStages = stageListener.getParentToChildren().get(stageId);
+      // set children field
+      List<Long> childrenStages =
+          stageLevelListener.getParentStageToChildrenStages().get(stageId);
       if (childrenStages != null) {
         List<Task> children = new ArrayList<>();
-        childrenStages.forEach(childrenStage -> children.addAll(stageToTasks.get(childrenStage)));
-        Long[] temp = children.stream().map(Task::getId).toArray(Long[]::new);
-        task.setChildren(ArrayUtils.toPrimitive(temp));
+        childrenStages.forEach(
+            childrenStage -> children.addAll(this.getStageToTasks().get(childrenStage)));
+        long[] childrenTaskIds = children.stream()
+            .map(Task::getId)
+            .mapToLong(Long::longValue)
+            .toArray();
+        task.setChildren(childrenTaskIds);
       }
 
-      // resource related fields
-      final int resourceProfileId = stageListener.getStageToResource().get(stageId);
+      // set resource related fields
+      final int resourceProfileId =
+          stageLevelListener.getStageToResource().getOrDefault(stageId, -1);
       final ResourceProfile resourceProfile =
           sparkContext.resourceProfileManager().resourceProfileFromId(resourceProfileId);
       final List<TaskResourceRequest> resources = JavaConverters.seqAsJavaList(