+++ adapter/spark/src/test/java/com/asml/apa/wta/spark/listener/ApplicationLevelListenerTest.java
@@ -4,7 +4,6 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
 
-import com.asml.apa.wta.core.model.Task;
 import com.asml.apa.wta.core.model.Workload;
 import java.util.Properties;
 import org.apache.spark.executor.ExecutorMetrics;
@@ -25,7 +24,7 @@ class ApplicationLevelListenerTest extends BaseLevelListenerTest {
 
   SparkListenerApplicationEnd applicationEndObj;
 
-  TaskInfo testTaskInfo;
+  TaskInfo testTaskInfo1;
 
   TaskInfo testTaskInfo2;
 
@@ -33,11 +32,11 @@ class ApplicationLevelListenerTest extends BaseLevelListenerTest {
 
   TaskInfo testTaskInfo4;
 
-  StageInfo testStageInfo;
+  StageInfo testStageInfo1;
 
   StageInfo testStageInfo2;
 
-  SparkListenerTaskEnd taskEndEvent;
+  SparkListenerTaskEnd taskEndEvent1;
 
   SparkListenerTaskEnd taskEndEvent2;
 
@@ -45,65 +44,88 @@ class ApplicationLevelListenerTest extends BaseLevelListenerTest {
 
   SparkListenerTaskEnd taskEndEvent4;
 
-  SparkListenerStageCompleted stageCompleted;
+  SparkListenerStageCompleted stageCompleted1;
 
   SparkListenerStageCompleted stageCompleted2;
 
+  int stageId1;
+
+  int stageId2;
+
+  long applicationDateEnd;
+
   @BeforeEach
   void setup() {
-    testTaskInfo = new TaskInfo(1, 0, 1, 50L, "testExecutor", "local", TaskLocality.NODE_LOCAL(), false);
-    testTaskInfo2 = new TaskInfo(2, 0, 1, 50L, "testExecutor", "local", TaskLocality.NODE_LOCAL(), false);
-    testTaskInfo3 = new TaskInfo(3, 0, 1, 50L, "testExecutor", "local", TaskLocality.NODE_LOCAL(), false);
-    testTaskInfo4 = new TaskInfo(4, 0, 1, 50L, "testExecutor", "local", TaskLocality.NODE_LOCAL(), false);
+    testTaskInfo1 = new TaskInfo(0, 0, 1, 50L, "testExecutor", "local", TaskLocality.NODE_LOCAL(), false);
+    testTaskInfo2 = new TaskInfo(1, 0, 1, 50L, "testExecutor", "local", TaskLocality.NODE_LOCAL(), false);
+    testTaskInfo3 = new TaskInfo(2, 0, 1, 50L, "testExecutor", "local", TaskLocality.NODE_LOCAL(), false);
+    testTaskInfo4 = new TaskInfo(3, 0, 1, 50L, "testExecutor", "local", TaskLocality.NODE_LOCAL(), false);
     ListBuffer<Object> parents = new ListBuffer<>();
-    TaskMetrics mockedMetrics = mock(TaskMetrics.class);
-    ShuffleWriteMetrics mockedShuffleMetrics = mock(ShuffleWriteMetrics.class);
-    when(mockedMetrics.executorRunTime()).thenReturn(100L);
-    when(mockedMetrics.diskBytesSpilled()).thenReturn(100L);
-    when(mockedMetrics.shuffleWriteMetrics()).thenReturn(mockedShuffleMetrics);
-    when(mockedShuffleMetrics.bytesWritten()).thenReturn(100L);
+
+    ShuffleWriteMetrics mockedShuffleMetrics1 = mock(ShuffleWriteMetrics.class);
+    ShuffleWriteMetrics mockedShuffleMetrics2 = mock(ShuffleWriteMetrics.class);
+    ShuffleWriteMetrics mockedShuffleMetrics3 = mock(ShuffleWriteMetrics.class);
+    ShuffleWriteMetrics mockedShuffleMetrics4 = mock(ShuffleWriteMetrics.class);
+    when(mockedShuffleMetrics1.bytesWritten()).thenReturn(100L);
+    when(mockedShuffleMetrics2.bytesWritten()).thenReturn(-1L);
+    when(mockedShuffleMetrics3.bytesWritten()).thenReturn(0L);
+    when(mockedShuffleMetrics4.bytesWritten()).thenReturn(50L);
+
+    TaskMetrics mockedMetrics1 = mock(TaskMetrics.class);
+    when(mockedMetrics1.peakExecutionMemory()).thenReturn(100L);
+    when(mockedMetrics1.executorRunTime()).thenReturn(100L);
+    when(mockedMetrics1.diskBytesSpilled()).thenReturn(100L);
+    when(mockedMetrics1.shuffleWriteMetrics()).thenReturn(mockedShuffleMetrics1);
 
     TaskMetrics mockedMetrics2 = mock(TaskMetrics.class);
-    ShuffleWriteMetrics mockedShuffleMetrics2 = mock(ShuffleWriteMetrics.class);
     when(mockedMetrics2.peakExecutionMemory()).thenReturn(-1L);
     when(mockedMetrics2.diskBytesSpilled()).thenReturn(-1L);
     when(mockedMetrics2.executorRunTime()).thenReturn(-1L);
     when(mockedMetrics2.shuffleWriteMetrics()).thenReturn(mockedShuffleMetrics2);
-    when(mockedShuffleMetrics2.bytesWritten()).thenReturn(-1L);
 
     TaskMetrics mockedMetrics3 = mock(TaskMetrics.class);
-    ShuffleWriteMetrics mockedShuffleMetrics3 = mock(ShuffleWriteMetrics.class);
     when(mockedMetrics3.peakExecutionMemory()).thenReturn(-1L);
-    when(mockedMetrics3.diskBytesSpilled()).thenReturn(1L);
+    when(mockedMetrics3.diskBytesSpilled()).thenReturn(0L);
     when(mockedMetrics3.executorRunTime()).thenReturn(-1L);
-    when(mockedMetrics3.shuffleWriteMetrics()).thenReturn(mockedShuffleMetrics2);
-    when(mockedShuffleMetrics3.bytesWritten()).thenReturn(0L);
+    when(mockedMetrics3.shuffleWriteMetrics()).thenReturn(mockedShuffleMetrics3);
 
-    testStageInfo =
-        new StageInfo(5, 0, "test", 50, null, new ListBuffer<>(), "None", mockedMetrics, null, null, 100);
-    parents.$plus$eq(5);
-    testStageInfo2 = new StageInfo(6, 0, "test", 50, null, parents, "None", mockedMetrics, null, null, 100);
-    taskEndEvent = new SparkListenerTaskEnd(
-        5, 1, "testTaskType", null, testTaskInfo, new ExecutorMetrics(), mockedMetrics);
+    TaskMetrics mockedMetrics4 = mock(TaskMetrics.class);
+    when(mockedMetrics4.peakExecutionMemory()).thenReturn(30L);
+    when(mockedMetrics4.diskBytesSpilled()).thenReturn(20L);
+    when(mockedMetrics4.executorRunTime()).thenReturn(80L);
+    when(mockedMetrics4.shuffleWriteMetrics()).thenReturn(mockedShuffleMetrics4);
+
+    stageId1 = 2;
+    stageId2 = 10;
+    parents.$plus$eq(stageId1);
+
+    testStageInfo1 = new StageInfo(
+        stageId1, 0, "test", 50, null, new ListBuffer<>(), "None", mockedMetrics1, null, null, 100);
+    testStageInfo2 = new StageInfo(stageId2, 0, "test", 50, null, parents, "None", mockedMetrics1, null, null, 100);
+
+    taskEndEvent1 = new SparkListenerTaskEnd(
+        stageId1, 1, "testTaskType", null, testTaskInfo1, new ExecutorMetrics(), mockedMetrics1);
     taskEndEvent2 = new SparkListenerTaskEnd(
-        5, 1, "testTaskType", null, testTaskInfo2, new ExecutorMetrics(), mockedMetrics);
+        stageId1, 1, "testTaskType", null, testTaskInfo2, new ExecutorMetrics(), mockedMetrics1);
     taskEndEvent3 = new SparkListenerTaskEnd(
-        6, 1, "testTaskType", null, testTaskInfo3, new ExecutorMetrics(), mockedMetrics2);
+        stageId2, 1, "testTaskType", null, testTaskInfo3, new ExecutorMetrics(), mockedMetrics2);
     taskEndEvent4 = new SparkListenerTaskEnd(
-        6, 1, "testTaskType", null, testTaskInfo4, new ExecutorMetrics(), mockedMetrics3);
-
-    stageCompleted = new SparkListenerStageCompleted(testStageInfo);
+        stageId2, 1, "testTaskType", null, testTaskInfo4, new ExecutorMetrics(), mockedMetrics3);
+    stageCompleted1 = new SparkListenerStageCompleted(testStageInfo1);
     stageCompleted2 = new SparkListenerStageCompleted(testStageInfo2);
-    applicationEndObj = new SparkListenerApplicationEnd(mockedSparkContext.startTime() + 1000L);
+
+    applicationDateEnd = mockedSparkContext.startTime() + 1000L;
+    applicationEndObj = new SparkListenerApplicationEnd(applicationDateEnd);
   }
 
   @Test
-  void doingNothingTest() {
-    ListBuffer<StageInfo> stageBuffer = new ListBuffer<>();
-    SparkListenerJobStart jobStart = new SparkListenerJobStart(1, 2L, stageBuffer.toList(), new Properties());
-    fakeApplicationListener.onApplicationEnd(applicationEndObj);
-    Workload workload = fakeApplicationListener.getProcessedObjects().get(0);
-    assertThat(fakeApplicationListener.getProcessedObjects().size()).isEqualTo(1);
+  void workloadBuiltWithDefaultMetricValues() {
+    assertThat(fakeApplicationListener1.getProcessedObjects()).hasSize(0);
+    fakeApplicationListener1.onApplicationEnd(applicationEndObj);
+    assertThat(fakeApplicationListener1.getProcessedObjects()).hasSize(1);
+
+    Workload workload = fakeApplicationListener1.getProcessedObjects().get(0);
+    assertThat(fakeApplicationListener1.getProcessedObjects().size()).isEqualTo(1);
     assertThat(workload.getFirstQuartileResourceTask()).isEqualTo(-1.0);
     assertThat(workload.getMaxResourceTask()).isEqualTo(-1.0);
     assertThat(workload.getCovResourceTask()).isEqualTo(-1.0);
@@ -123,77 +145,80 @@ class ApplicationLevelListenerTest extends BaseLevelListenerTest {
   }
 
   @Test
-  void uninitializedTest() {
-    ListBuffer<StageInfo> stageBuffer = new ListBuffer<>();
-    stageBuffer.$plus$eq(testStageInfo2);
-    SparkListenerJobStart jobStart = new SparkListenerJobStart(1, 2L, stageBuffer.toList(), new Properties());
-    fakeTaskListener.onJobStart(jobStart);
-    fakeStageListener.onJobStart(jobStart);
-    fakeTaskListener.onTaskEnd(taskEndEvent3);
-    fakeTaskListener.onTaskEnd(taskEndEvent3);
-    fakeStageListener.onStageCompleted(stageCompleted2);
-    fakeApplicationListener.onApplicationEnd(applicationEndObj);
-    Workload workload = fakeApplicationListener.getProcessedObjects().get(0);
-    assertThat(fakeApplicationListener.getProcessedObjects().size()).isEqualTo(1);
-    assertThat(workload.getFirstQuartileResourceTask()).isEqualTo(20);
-    assertThat(workload.getMaxResourceTask()).isEqualTo(20);
-    assertThat(workload.getCovResourceTask()).isEqualTo(0);
-    assertThat(workload.getCovDiskSpaceUsage()).isEqualTo(-1);
-    assertThat(workload.getFirstQuartileDiskSpaceUsage()).isEqualTo(-1);
-    assertThat(workload.getMinDiskSpaceUsage()).isEqualTo(-1);
-    assertThat(workload.getStdMemory()).isEqualTo(-1);
-    assertThat(workload.getMedianMemory()).isEqualTo(-1);
-    assertThat(workload.getMinMemory()).isEqualTo(-1);
-    assertThat(workload.getTotalTasks()).isEqualTo(0);
-    assertThat(workload.getMeanEnergy()).isEqualTo(-1.0);
-    assertThat(workload.getMeanMemory()).isEqualTo(-1.0);
-    assertThat(workload.getMeanResourceTask()).isEqualTo(20);
-    assertThat(workload.getMeanNetworkUsage()).isEqualTo(-1.0);
-    assertThat(workload.getMeanDiskSpaceUsage()).isEqualTo(-1.0);
-    assertThat(workload.getTotalResourceSeconds()).isEqualTo(-1);
+  void workloadGeneralMetricsCollected() {
+    assertThat(fakeApplicationListener1.getProcessedObjects()).hasSize(0);
+    fakeApplicationListener1.onApplicationEnd(applicationEndObj);
+    assertThat(fakeApplicationListener1.getProcessedObjects()).hasSize(1);
+
+    Workload workload = fakeApplicationListener1.getProcessedObjects().get(0);
+    assertThat(workload.getDomain()).isEqualTo(fakeConfig1.getDomain());
+    assertThat(workload.getAuthors()).isEqualTo(fakeConfig1.getAuthors());
+    assertThat(workload.getWorkloadDescription()).isEqualTo(fakeConfig1.getDescription());
+    long sutStartTime = mockedSparkContext.startTime();
+    assertThat(workload.getDateStart()).isEqualTo(sutStartTime);
+    assertThat(workload.getDateEnd()).isEqualTo(applicationDateEnd);
   }
 
   @Test
-  void zeroMeanTest() {
-    ListBuffer<StageInfo> stageBuffer = new ListBuffer<>();
-    stageBuffer.$plus$eq(testStageInfo2);
-    SparkListenerJobStart jobStart = new SparkListenerJobStart(1, 2L, stageBuffer.toList(), new Properties());
-    fakeTaskListener.onJobStart(jobStart);
-    fakeStageListener.onJobStart(jobStart);
-    fakeTaskListener.onTaskEnd(taskEndEvent4);
-    fakeStageListener.onStageCompleted(stageCompleted2);
-    fakeApplicationListener.onApplicationEnd(applicationEndObj);
-    Workload workload = fakeApplicationListener.getProcessedObjects().get(0);
-    assertThat(fakeApplicationListener.getProcessedObjects().size()).isEqualTo(1);
-    assertThat(workload.getFirstQuartileResourceTask()).isEqualTo(20);
-    assertThat(workload.getMaxResourceTask()).isEqualTo(20);
-    assertThat(workload.getCovResourceTask()).isEqualTo(0);
-    assertThat(workload.getCovDiskSpaceUsage()).isEqualTo(-1);
-    assertThat(workload.getFirstQuartileDiskSpaceUsage()).isEqualTo(0);
-    assertThat(workload.getMinDiskSpaceUsage()).isEqualTo(0);
-    assertThat(workload.getStdMemory()).isEqualTo(-1);
-    assertThat(workload.getMedianMemory()).isEqualTo(-1);
-    assertThat(workload.getMinMemory()).isEqualTo(-1);
+  void workloadCountMetricsCollected() {
+    assertThat(fakeApplicationListener1.getProcessedObjects()).hasSize(0);
+    fakeApplicationListener1.onApplicationEnd(applicationEndObj);
+    assertThat(fakeApplicationListener1.getProcessedObjects()).hasSize(1);
+
+    Workload workload = fakeApplicationListener1.getProcessedObjects().get(0);
+    assertThat(workload.getTotalWorkflows()).isEqualTo(0);
     assertThat(workload.getTotalTasks()).isEqualTo(0);
+    assertThat(workload.getNumUsers()).isEqualTo(0);
+    assertThat(workload.getNumGroups()).isEqualTo(0);
+    assertThat(workload.getNumSites()).isEqualTo(0);
+    assertThat(workload.getNumResources()).isEqualTo(-1L);
+    assertThat(workload.getTotalResourceSeconds()).isEqualTo(-1L);
+  }
+
+  @Test
+  void workloadResourceMetricsCollected() {
+    ListBuffer<StageInfo> stageBuffer = new ListBuffer<>();
+    stageBuffer.$plus$eq(testStageInfo1);
+    stageBuffer.$plus$eq(testStageInfo2);
+    SparkListenerJobStart jobStart = new SparkListenerJobStart(0, 2L, stageBuffer.toList(), new Properties());
+    fakeTaskListener1.onJobStart(jobStart);
+    fakeStageListener1.onJobStart(jobStart);
+    fakeTaskListener1.onTaskEnd(taskEndEvent1);
+    fakeTaskListener1.onTaskEnd(taskEndEvent2);
+    fakeStageListener1.onStageCompleted(stageCompleted1);
+    fakeTaskListener1.onTaskEnd(taskEndEvent3);
+    fakeTaskListener1.onTaskEnd(taskEndEvent4);
+    fakeStageListener1.onStageCompleted(stageCompleted2);
+
+    assertThat(fakeApplicationListener1.getProcessedObjects()).hasSize(0);
+    fakeApplicationListener1.onApplicationEnd(applicationEndObj);
+    assertThat(fakeApplicationListener1.getProcessedObjects()).hasSize(1);
+
+    Workload workload = fakeApplicationListener1.getProcessedObjects().get(0);
+    assertThat(fakeApplicationListener1.getProcessedObjects().size()).isEqualTo(1);
     assertThat(workload.getMeanEnergy()).isEqualTo(-1.0);
     assertThat(workload.getMeanMemory()).isEqualTo(-1.0);
-    assertThat(workload.getMeanResourceTask()).isEqualTo(20);
-    assertThat(workload.getMeanNetworkUsage()).isEqualTo(-1);
-    assertThat(workload.getMeanDiskSpaceUsage()).isEqualTo(0);
-    assertThat(workload.getTotalResourceSeconds()).isEqualTo(-1);
+    assertThat(workload.getMeanResourceTask()).isEqualTo(-1.0);
+    assertThat(workload.getMeanNetworkUsage()).isEqualTo(-1.0);
+    assertThat(workload.getMeanDiskSpaceUsage()).isEqualTo(133.33333333333334);
+    assertThat(workload.getTotalResourceSeconds()).isEqualTo(-1.0);
   }
 
   @Test
   void stageWithoutTaskTest() {
     ListBuffer<StageInfo> stageBuffer = new ListBuffer<>();
     stageBuffer.$plus$eq(testStageInfo2);
-    SparkListenerJobStart jobStart = new SparkListenerJobStart(1, 2L, stageBuffer.toList(), new Properties());
-    fakeTaskListener.onJobStart(jobStart);
-    fakeStageListener.onJobStart(jobStart);
-    fakeStageListener.onStageCompleted(stageCompleted2);
-    fakeApplicationListener.onApplicationEnd(applicationEndObj);
-    Workload workload = fakeApplicationListener.getProcessedObjects().get(0);
-    assertThat(fakeApplicationListener.getProcessedObjects().size()).isEqualTo(1);
+    SparkListenerJobStart jobStart = new SparkListenerJobStart(0, 2L, stageBuffer.toList(), new Properties());
+    fakeTaskListener1.onJobStart(jobStart);
+    fakeStageListener1.onJobStart(jobStart);
+    fakeStageListener1.onStageCompleted(stageCompleted2);
+
+    assertThat(fakeApplicationListener1.getProcessedObjects()).hasSize(0);
+    fakeApplicationListener1.onApplicationEnd(applicationEndObj);
+    assertThat(fakeApplicationListener1.getProcessedObjects()).hasSize(1);
+
+    Workload workload = fakeApplicationListener1.getProcessedObjects().get(0);
+    assertThat(fakeApplicationListener1.getProcessedObjects().size()).isEqualTo(1);
     assertThat(workload.getFirstQuartileResourceTask()).isEqualTo(-1);
     assertThat(workload.getMaxResourceTask()).isEqualTo(-1);
     assertThat(workload.getCovResourceTask()).isEqualTo(-1);
@@ -213,140 +238,29 @@ class ApplicationLevelListenerTest extends BaseLevelListenerTest {
   }
 
   @Test
-  void applicationEndTwoTimesTest() {
+  void applicationEndEnteringForbiddenBranchWillGracefullyTerminate() {
     ListBuffer<StageInfo> stageBuffer = new ListBuffer<>();
+    stageBuffer.$plus$eq(testStageInfo1);
     stageBuffer.$plus$eq(testStageInfo2);
-    SparkListenerJobStart jobStart = new SparkListenerJobStart(1, 2L, stageBuffer.toList(), new Properties());
-    fakeTaskListener.onJobStart(jobStart);
-    fakeStageListener.onJobStart(jobStart);
-    fakeTaskListener.onTaskEnd(taskEndEvent3);
-    fakeTaskListener.onTaskEnd(taskEndEvent3);
-    fakeStageListener.onStageCompleted(stageCompleted2);
-    fakeApplicationListener.onApplicationEnd(applicationEndObj);
-    fakeApplicationListener.onApplicationEnd(applicationEndObj);
-    Workload workload = fakeApplicationListener.getProcessedObjects().get(0);
-    assertThat(fakeApplicationListener.getProcessedObjects().size()).isEqualTo(1);
-    assertThat(workload.getFirstQuartileResourceTask()).isEqualTo(20);
-    assertThat(workload.getMaxResourceTask()).isEqualTo(20);
-    assertThat(workload.getCovResourceTask()).isEqualTo(0);
-    assertThat(workload.getCovDiskSpaceUsage()).isEqualTo(-1);
-    assertThat(workload.getFirstQuartileDiskSpaceUsage()).isEqualTo(-1);
-    assertThat(workload.getMinDiskSpaceUsage()).isEqualTo(-1);
-    assertThat(workload.getStdMemory()).isEqualTo(-1);
-    assertThat(workload.getMedianMemory()).isEqualTo(-1);
-    assertThat(workload.getMinMemory()).isEqualTo(-1);
-    assertThat(workload.getTotalTasks()).isEqualTo(0);
-    assertThat(workload.getMeanEnergy()).isEqualTo(-1.0);
-    assertThat(workload.getMeanMemory()).isEqualTo(-1.0);
-    assertThat(workload.getMeanResourceTask()).isEqualTo(20);
-    assertThat(workload.getMeanNetworkUsage()).isEqualTo(-1.0);
-    assertThat(workload.getMeanDiskSpaceUsage()).isEqualTo(-1.0);
-    assertThat(workload.getTotalResourceSeconds()).isEqualTo(-1);
-  }
+    SparkListenerJobStart jobStart1 = new SparkListenerJobStart(0, 2L, stageBuffer.toList(), new Properties());
+    SparkListenerJobStart jobStart2 = new SparkListenerJobStart(1, 2L, stageBuffer.toList(), new Properties());
 
-  @Test
-  void parentChildrenAggregationTest() {
-    ListBuffer<StageInfo> stageBuffer = new ListBuffer<>();
-    stageBuffer.$plus$eq(testStageInfo);
-    stageBuffer.$plus$eq(testStageInfo2);
-    SparkListenerJobStart jobStart = new SparkListenerJobStart(1, 2L, stageBuffer.toList(), new Properties());
-    fakeTaskListener.onJobStart(jobStart);
-    fakeStageListener.onJobStart(jobStart);
-    fakeTaskListener.onTaskEnd(taskEndEvent);
-    fakeTaskListener.onTaskEnd(taskEndEvent2);
-    fakeStageListener.onStageCompleted(stageCompleted);
-    fakeTaskListener.onTaskEnd(taskEndEvent3);
-    fakeTaskListener.onTaskEnd(taskEndEvent4);
-    fakeStageListener.onStageCompleted(stageCompleted2);
-    fakeApplicationListener.onApplicationEnd(applicationEndObj);
-    Task task = fakeTaskListener.getProcessedObjects().get(0);
-    assertThat(task.getParents().length).isEqualTo(0);
-    assertThat(task.getChildren().length).isEqualTo(2);
-    assertThat(task.getChildren()).contains(4, 5);
-    task = fakeTaskListener.getProcessedObjects().get(1);
-    assertThat(task.getParents().length).isEqualTo(0);
-    assertThat(task.getChildren().length).isEqualTo(2);
-    assertThat(task.getChildren()).contains(4, 5);
-    task = fakeTaskListener.getProcessedObjects().get(2);
-    assertThat(task.getParents().length).isEqualTo(2);
-    assertThat(task.getParents()).contains(2, 3);
-    assertThat(task.getChildren().length).isEqualTo(0);
-    task = fakeTaskListener.getProcessedObjects().get(3);
-    assertThat(task.getParents().length).isEqualTo(2);
-    assertThat(task.getParents()).contains(2, 3);
-    assertThat(task.getChildren().length).isEqualTo(0);
-    Workload workload = fakeApplicationListener.getProcessedObjects().get(0);
-    assertThat(fakeApplicationListener.getProcessedObjects().size()).isEqualTo(1);
-    assertThat(workload.getFirstQuartileResourceTask()).isEqualTo(20);
-    assertThat(workload.getMaxResourceTask()).isEqualTo(20);
-    assertThat(workload.getCovResourceTask()).isEqualTo(0);
-    assertThat(workload.getCovDiskSpaceUsage()).isEqualTo(0.7071067811865474);
-    assertThat(workload.getFirstQuartileDiskSpaceUsage()).isEqualTo(0);
-    assertThat(workload.getMinDiskSpaceUsage()).isEqualTo(0);
-    assertThat(workload.getStdMemory()).isEqualTo(-1);
-    assertThat(workload.getMedianMemory()).isEqualTo(-1);
-    assertThat(workload.getMinMemory()).isEqualTo(-1);
-    assertThat(workload.getTotalTasks()).isEqualTo(0);
-    assertThat(workload.getMeanEnergy()).isEqualTo(-1.0);
-    assertThat(workload.getMeanMemory()).isEqualTo(-1.0);
-    assertThat(workload.getMeanResourceTask()).isEqualTo(20);
-    assertThat(workload.getMeanNetworkUsage()).isEqualTo(-1.0);
-    assertThat(workload.getMeanDiskSpaceUsage()).isEqualTo(133.33333333333334);
-    assertThat(workload.getTotalResourceSeconds()).isEqualTo(4000);
-  }
+    fakeTaskListener1.onJobStart(jobStart1);
+    fakeStageListener1.onJobStart(jobStart1);
+    fakeTaskListener1.onTaskEnd(taskEndEvent1);
+    fakeTaskListener1.onTaskEnd(taskEndEvent2);
+    fakeStageListener1.onStageCompleted(stageCompleted1);
+    assertThat(fakeApplicationListener1.getProcessedObjects()).hasSize(0);
+    fakeApplicationListener1.onApplicationEnd(applicationEndObj);
+    assertThat(fakeApplicationListener1.getProcessedObjects()).hasSize(1);
 
-  @Test
-  void parentChildrenAggregationAltTest() {
-    ListBuffer<StageInfo> stageBuffer = new ListBuffer<>();
-    stageBuffer.$plus$eq(testStageInfo);
-    stageBuffer.$plus$eq(testStageInfo2);
-    SparkListenerJobStart jobStart = new SparkListenerJobStart(1, 2L, stageBuffer.toList(), new Properties());
-    fakeTaskListener2.onJobStart(jobStart);
-    fakeStageListener2.onJobStart(jobStart);
-    fakeTaskListener2.onTaskEnd(taskEndEvent);
-    fakeTaskListener2.onTaskEnd(taskEndEvent2);
-    fakeStageListener2.onStageCompleted(stageCompleted);
-    fakeTaskListener2.onTaskEnd(taskEndEvent3);
-    fakeTaskListener2.onTaskEnd(taskEndEvent4);
-    fakeStageListener2.onStageCompleted(stageCompleted2);
-    fakeApplicationListener2.onApplicationEnd(applicationEndObj);
-    Workload workload = fakeApplicationListener2.getProcessedObjects().get(0);
-    assertThat(fakeApplicationListener2.getProcessedObjects().size()).isEqualTo(1);
-    assertThat(workload.getFirstQuartileResourceTask()).isEqualTo(-1);
-    assertThat(workload.getMaxResourceTask()).isEqualTo(-1);
-    assertThat(workload.getCovResourceTask()).isEqualTo(-1);
-    assertThat(workload.getMeanResourceTask()).isEqualTo(-1);
-    assertThat(workload.getTotalResourceSeconds()).isEqualTo(-1);
-  }
-
-  @Test
-  void applicationListenerCollectsDesiredInformation() {
-    fakeApplicationListener.onApplicationEnd(applicationEndObj);
-    assertThat(fakeApplicationListener.getProcessedObjects()).hasSize(1);
-
-    Workload workload = fakeApplicationListener.getProcessedObjects().get(0);
-    assertThat(workload.getTotalTasks()).isEqualTo(0);
-    assertThat(workload.getDomain()).isEqualTo(fakeConfig.getDomain());
-    long sutStartTime = mockedSparkContext.startTime();
-    assertThat(workload.getDateStart()).isEqualTo(sutStartTime);
-    assertThat(workload.getDateEnd()).isEqualTo(sutStartTime + 1000L);
-
-    assertThat(fakeApplicationListener.getProcessedObjects().size()).isEqualTo(1);
-    assertThat(workload.getFirstQuartileResourceTask()).isEqualTo(-1.0);
-    assertThat(workload.getMaxResourceTask()).isEqualTo(-1.0);
-    assertThat(workload.getCovResourceTask()).isEqualTo(-1.0);
-    assertThat(workload.getCovDiskSpaceUsage()).isEqualTo(-1);
-    assertThat(workload.getFirstQuartileDiskSpaceUsage()).isEqualTo(-1);
-    assertThat(workload.getMinDiskSpaceUsage()).isEqualTo(-1);
-    assertThat(workload.getStdMemory()).isEqualTo(-1);
-    assertThat(workload.getMedianMemory()).isEqualTo(-1);
-    assertThat(workload.getMinMemory()).isEqualTo(-1);
-    assertThat(workload.getTotalTasks()).isEqualTo(0);
-    assertThat(workload.getMeanEnergy()).isEqualTo(-1.0);
-    assertThat(workload.getMeanMemory()).isEqualTo(-1.0);
-    assertThat(workload.getMeanResourceTask()).isEqualTo(-1.0);
-    assertThat(workload.getMeanNetworkUsage()).isEqualTo(-1.0);
-    assertThat(workload.getMeanDiskSpaceUsage()).isEqualTo(-1.0);
-    assertThat(workload.getTotalResourceSeconds()).isEqualTo(-1);
+    fakeTaskListener1.onJobStart(jobStart2);
+    fakeStageListener1.onJobStart(jobStart2);
+    fakeTaskListener1.onTaskEnd(taskEndEvent3);
+    fakeTaskListener1.onTaskEnd(taskEndEvent4);
+    fakeStageListener1.onStageCompleted(stageCompleted2);
+    assertThat(fakeApplicationListener1.getProcessedObjects()).hasSize(1);
+    fakeApplicationListener1.onApplicationEnd(applicationEndObj);
+    assertThat(fakeApplicationListener1.getProcessedObjects()).hasSize(1);
   }
 }